+++ src/ByteDev.Cmd/HorizontalLineFactory.cs
@@ -14,7 +14,7 @@ namespace ByteDev.Cmd
 
         public static string Create(Table table)
         {
-            var totalValueWidths = table.Columns * table.GetLongestElementLength();
+            var totalValueWidths = table.Columns * table.GetLongestCellValueLength();
 
             var totalPaddingsLength = (table.LeftPadding.Length + table.RightPadding.Length) * table.Columns;