+++ src/ByteDev.Cmd/Tables/CellValueAlignment.cs
@@ -1,0 +1,8 @@
+﻿namespace ByteDev.Cmd.Tables
+{
+    public enum CellValueAlignment
+    {
+        Left,
+        Right
+    }
+}